Add fix in preprocess_qual_conditions
Because it is incorrect to replace SubLinks in the select list or
basically anywhere outside the qual, put the check in
preprocess_qual_conditions. If the SubLink type is ANY (SubLink type ALL
would fail because if relation is empty, then it would return true and
I have to think more about EXPR type) and, after constant folding on the
testexpr it is a Const and the const is NULL, replace the SubLink with
its testexpr because the testexpr will never be true. This way, the
SubLink is not replaced with a subplan later.