hashloop fallback
First part is to "chunk" the inner file into arbitrary partitions of
work_mem size

This chunks inner file and makes it so that the offset is along tuple
bounds.

Note that this makes it impossible to increase nbatches during the
loading of batches after initial hashtable creation

In preparation for doing this chunking, separate advance batch and load
batch. advance batch only if page offset is reset to 0, then load that
part of the batch

Second part was to: implement outer tuple batch rewinding per chunk of
inner batch

Would be a simple rewind and replay of outer side for each chunk of
inner if it weren't for LOJ.
Because we need to wait to emit NULL-extended tuples for LOJ until after
all chunks of inner have been processed.

To do this without incurring additional memory pressure, use a temporary
Buffile to capture the match status of each outer side tuple. Use one
bit per tuple to represent the match status, and, since for
parallel-oblivious hashjoin the outer side tuples are encountered in a
deterministic order, synchronizing the outer tuples match status file
with the outer tuples in the batch file to decide which ones to emit
NULL-extended is easy and can be done with a simple counter.

For non-hashloop fallback scenario (including batch 0), this file is not
created and unmatched outer tuples should be emitted as they are
encountered.

OuterTupleMatchStatuses are in a file as a bitmap instead of in memory