Implement outer tuple batch rewinding per chunk of inner batch
Would be a simple rewind and replay of outer side for each chunk of
inner if it weren't for LOJ.
Because we need to wait to emit NULL-extended tuples for LOJ until after
all chunks of inner have been processed.
To do this, make a list with an entry for each outer tuple and keep
track of its match status. Also, keep track of its offset so that we can
access the file at that offset in case the tuplesa re not processed in
order (like in parallel case)

After we have processed all chunks of inner we can loop through the
outer tuple list and and look up all unmatched tuples and NULL extend
and emit them

Logic for creating list entries for outer batch tuples in
HJ_NEED_NEW_OUTER:

new object o
current cur
head head

if node->first_chunk is true
    need_new_outer
	make object o
	if head is NULL
	    head = o
	    cur = o
	else
	    cur.next = o
	    cur = o
    load_inner_batch
	cur = NULL
	head = NULL
if node->first_chunk is false
    need_new_outer
	if cur is NULL
	    cur = HEAD
	else
	    cur = cur.next
    load_inner_batch
	cur = NULL