Last worker prints tuplenums for all workers outer batch files
Add more logging
Last worker should be able to print the tuplenums of all the read_files
of the workers and print out their tuples

print out all of the sharedsegmentnames and buffilenames and how many
tuples in each for a given batch -- maybe also print the tuple values
this will tell me, does each backend have a read_file which is different (also
check write_file) in terms of the tuples it contains
if so, by looping through participants, will I be able to get to all of them?
is it each sharedsegment that is different or each read_file or what?

it is still unclear to me if the read_files per worker for a batch will
have the same content

next step, use logging and debugging to determine if the content of the
different outerbatch files for a sharedfileset contain the same or
different tuples
then determine if we can loop through all of them and do it with the
last worker.