Lots of hacks to add chunk numbers
On the inner side, we need to number the chunks when we run out of
work_mem. Though this should be more in sync with when we disable
growth, for now, I am simply incrementing the chunk number when I exceed
work_mem for a given batch and resetting this chunk num counter when I
resize the number of batches and need to repartition.

This chunk number will be potentially greater than 1 even when we
haven't decided to fall back.

The next step is to use this chunk number when inserting into the inner
side hashtable if we are falling back to block nested hashloop join and
keep track of the current chunk and then rescan the outer.