Fix chunking logic to only do it if parallel_hashloop_fallback is true
For now, only increment chunk number if parallel_hashloop_fallback is
true -- this is only set to true when we have an "excessive" number of
batch increases (currently 2).
It is arguable that this is not the best way to do this, since, if we
have exceeded work_mem, maybe we should use the fallback method.
However, with the current state of the code this was easier. It also
felt less aggressive.