WIP broken
We tried adding a walker to find all of the columns in targetlist and
qual which are part of an aggregate so that we ensure that we write both
those and the grouping key to the spill file. The code is bad because we
hard-coded it to handle when there is a cast (FUNCEXPR) on the variable
in the agg.

This still doesn't work, it produces wrong results, but, at least we
aren't NULLing out the wrong columns.

set enable_hashagg = true;
set work_mem=64;
select g%100000 as c1, sum(g::numeric) as c2, count(*) as c3
  from generate_series(0, 199999) g
  group by g%100000 order by c2 desc limit 3;