Implement Adaptive Hashjoin
Serial Hashloop Fallback:

"Chunk" the inner file into arbitrary partitions of work_mem size
offset along tuple bounds while loading the batch into the hashtable.

Note that this makes it impossible to increase nbatches during the
loading of batches after initial hashtable creation.

In preparation for doing this chunking, separate "advance batch" and
"load batch".

Implement outer tuple batch rewinding per chunk of inner batch. Would
be a simple rewind and replay of outer side for each chunk of inner if
it weren't for LOJ. Because we need to wait to emit NULL-extended
tuples for LOJ until after all chunks of inner have been processed.

To do this without incurring additional memory pressure, use a
temporary BufFile to capture the match status of each outer side
tuple. Use one bit per tuple to represent the match status, and, since
for parallel-oblivious hashjoin the outer side tuples are encountered
in a deterministic order, synchronizing the outer tuples match status
file with the outer tuples in the batch file to decide which ones to
emit NULL-extended is easy and can be done with a simple counter.

For non-hashloop fallback scenario (including batch 0), this file is
not created and unmatched outer tuples should be emitted as they are
encountered.

Parallel Hashloop Fallback:

During initial allocation of the hashtable, each time the number of
batches is increased, a new variable in the ParallelHashJoinState,
batch_increases, is incremented.

In PHJ_GROW_BATCHES_DECIDING, if pstate->batch_increases >= 2,
parallel_hashloop_fallback will be enabled for qualifying batches.
From then on, if a batch is still too large to fit into the
space_allowed, then parallel_hashloop_fallback is set on that batch.
It will not be allowed to divide further and, during execution, the
fallback strategy will be used.

For a batch which has parallel_hashloop_fallback set, tuples inserted
into the the batch's inner and outer batch files will have an
additional piece of metadata (other than the hashvalue). For the inner
side, this additional metadata is the chunk number, For the outer
side, this additional metadata is the tuple identifier--needed when
rescanning the outer side batch file for each chunk of the inner.

During execution of a parallel hashjoin batch which needs to fall
back, the worker will create an "outer match status file" which
contains a bitmap tracking which outer tuples have matched an inner
tuple. All bits in the worker's outer match status file are initially
unset. During probing, the worker will set the corresponding bit (the
bit at the index of the tuple identifier) in the outer match status
bitmap for an outer tuple which matches any inner tuple.

Workers probing a fallback batch will wait until all workers have
finished probing before moving on so that an elected worker can read
and combine the outer match status files into a single bitmap and use
it to emit unmatched outer tuples after all chunks of the inner side
have been processed.