Address barrier wait deadlock hazard
Previously, in the chunk phase machine in
ExecParallelHashJoinNewChunk(), we reused one chunk barrier for all
chunks and looped through the phases then set the phase back to the
initial phase and explicitly jumped there.
Now, we initialize an array of chunk barriers, one per chunk, and use a
new chunk barrier for each chunk.
After finishing probing a chunk, upon re-entering
ExecParallelHashJoinNewChunk(), workers will wait on the chunk barrier
for all participants to arrive.
This is okay because the barrier is advanced to the final phase as part
of this wait (per the comment in nodeHashJoin.c about deadlock risk with
waiting on barriers after emitting tuples).
The last worker to arrive will increment the chunk number.
All workers detach from the chunk barrier they are attached to and
select the next chunk barrier.

The hashtable is now reset in the first phase of the chunk phase machine
PHJ_CHUNK_ELECTING. Note that this will cause an unnecessary hashtable
reset for the first chunk.

The loading and probing phases of the chunk phase machine stay the same.

If a worker joins in the PHJ_CHUNK_DONE phase, it will simply detach
from the chunk barrier and move on to the next chunk barrier in the
array of chunk barriers.

In order to mitigate the other cause of deadlock hazard (workers wait on
the batch barrier after emitting tuples), now, in
ExecParallelHashJoinNewBatch(), if we are attached to a batch barrier
and it is a fallback batch, all workers will detach from the batch
barrier and then end their scan of that batch.
The last worker to detach will combine the outer match status files,
then it will detach from the batch, clean up the hashtable, and end its
scan of the inner side.
Then it will return and proceed to emit unmatched outer tuples.
In PHJ_BATCH_ELECTING, the worker that ends up allocating the hashtable
will also initialize the chunk barriers.

Also, this commit moves combined_bitmap to batch from hjstate. It will
be moved into the SharedBits store once that API is added.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>