Add SharedBits API
Add SharedBits API--a way for workers to collaboratively make a bitmap.
The SharedBits store is currently meant for each backend to write to its
own bitmap file in one phase and for a single worker to combine all of
the bitmaps into a combined bitmap in another phase. In other words, it
supports parallel write but not parallel scan (and not concurrent
read/write). This could be modified in the future.

Also, the SharedBits uses a SharedFileset which uses BufFiles. This is
not the ideal API for the bitmap. The access pattern is small sequential
writes and random reads. It would also be nice to maintain the fixed
size buffer but have an API that let us write an arbitrary number of
bytes to it in bufsize chunks without incurring additional function call
overhead.

This commit also moves the outer match status file and combined_bitmap
into a new SharedBits store.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>