Avoid rescanning inner tuples per stripe
Instead of reinitializing the SharedTuplestore for the inner side for
each stripe during fallback, each participant's read_page is set to the
begininng of the SharedTuplestoreChunk which contains the end of one
stripe and the beginning of another.

Previously all inner tuples were scanned and only tuples from the
current stripe were loaded.

Each SharedTuplestoreAccessor now has a variable start_page, which is
initialized when it is assigned its read_page (which will always be the
beginning of a SharedTuplestoreChunk).

While loading tuples into the hashtable, if a tuple is from a past
stripe, the worker skips it (that will happen when a stripe straddles
two SharedTuplestoreChunks). If a tuple is from the future, the worker
backs that SharedTuplestoreChunk out and sets the shared read_page (in
the shared SharedTuplestoreParticipant) back to its start_page.

There are a couple mechanisms to provide for synchronization that
address specific race conditions/synchronization points:

Scenario 1:
- given a batch which has multiple stripes and has chosen the fallback
  strategy, two workers have each started reading from a single participant file
- worker0 is assigned pages 0-4
- worker1 is assigned pages 4-8
- stripe 1 starts on page 0 and ends on page 3
- worker0 sees that a tuple on page 3 is from stripe 2, so it proceeds
  to back out the read_page for this participant from 8 to its
  start_page of 0
- read_page is now 0
- worker1 was descheduled or distracted and starts reading a bit later.
  It sees that the very first tuple on page 4 is from a future
  stripe, so it wants to back out read_page to its start_page: 4
- If worker1 was allowed to do this, read_page would incorrectly be 4
  and tuples from stripe 2 on pages 3 and 4 would not be loaded into the
  hashtable
To handle this, a worker can only set read_page to a start_page which
is less than the current value of read_page

Scenario 2:
- given a batch which has multiple stripes and has chosen the fallback
  strategy, worker0 reads from participant file 0 and worker1 reads from
  participant file 1
- worker0 is assigned pages 0-4 in file1
- stripe 1 starts on page 0 and ends on page 3
- the current stripe is stripe 1
- worker0 sees on page3 that stripe 2 starts, so it backs out the
  read_page to 0
- worker1 finishes participant file 1 and proceeds to read from
  participant file 0
- worker1 opens the file and goes to get read_page
- read_page is 0, so worker1 loads tuples from stripe 1 in pages 0-3
- now both workers have loaded the same tuples into the hashtable
To handle this scenario, the participant has a rewound flag, which
indicates if this participant has been rewound during loading of the
current stripe. If it has, a worker cannot be assigned a
SharedTuplestoreChunk. This flag is reset later.

In this patch, Hashjoin makes an unacceptable intrusion into the
SharedTuplestore API. I am looking for feedback on how to solve this.
Basically, because the SharedTuplestore does not know about stripes or
about HashJoin, the logic to decide if a tuple should be loaded into a
hashtable or not is in the stripe phase machine where tuples are loaded
into the hashtable. So, to ensure that workers have read from all
participant files before assuming all tuples from a stripe are
loaded, I have duplicated the logic from sts_parallel_scan_next() which
has workers try the next participant in the body of the tuple loading
loop in the stripe phase machine (see sts_ready_for_next_stripe() and
sts_seen_all_participants()).

This clearly needs to be fixed and it is arguable that there are other
intrusions into the SharedTuplestore API in these patches.

One option is to write each stripe for each participant to a different
file, preserving the idea that a worker is done with a read_file when it
is at EOF.