Detach unless last worker
One potential solution to the deadlock hazard where workers wait on the
stripe_barrier after emitting tuples is to restrict participation to a
single worker after the first stripe.

When returning after probing, a worker checks if it is the only worker
attached to the stripe_barrier. If it is not, it will detach
unconditionally and go look for another batch. If it is the only worker,
it will proceed with its work.

Also, if a worker attaches to the stripe_barrier any time after the
first stripe, it will detach and go look for another batch.