Validate hashtable and STS BUILD Debugging
This commit introduces validation logic (meant for debugging during
development only) to check that all of the tuples pulled out of child
nodes for inner and outer make it into the SharedTuplestores and hash
table.

The idea was to make a tool for the developer to isolate the source of
bugs by narrowing it down to either build or load/probe.

Because MinimalTuples have no unique identifier, this will only work for
tables with a first column that is an int that uniquely identifies a
tuple.

The usage of the validation code is behind the HJDEBUG macro. There is a
test case in the regress test join_hash.sql which would exercise the
validation code, however, all other tests which use parallel hash join
will crash with HJDEBUG defined.

So, instead we added a guc, phj_check. We would need to add an alternate
answer file with the guc on -- which is not done here.

Currently something about the validation functions causes wrong results.

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>