Alternative implementation using list
DOES NOT WORK
This is the outline of how using a single freelist-esque list of
TBMIterateResult in the actual BitmapHeapScanState node would work. It
doesn't work because I don't have access to the BitmapHeapScanState node
inside of heapam_scan_bitmap_next_block() in order to put the
TBMIterateResult back on the list once I have used it.
Also, the list API doesn't have a function (as far as I can tell) which
will remove a target list cell from the list without freeing the memory.