aio: Add asynchronous IO infrastructure.
Add a new GUC to control which AIO implementation should be used.

Provide io_method=worker as a default that is always available, using
auxiliary processes to perform asynchronous I/O using traditional synchronous
interfaces.

Provide io_method=io_uring only if built with --with-liburing, on linux 5.1+.

Support io_Method=posix_aio only if built with --with-posix-aio, on platforms
like FreeBSD and OSX. It builds and runs on Linux, but due to glibc'
implementation with worker threads, it may not perform well.

This itself only provides infrastructure, which later commits utilize to
actually use AIO.

XXX: Several pieces here should be split out, but are kept as part of
this commit for now, to make it easier to revise APIs etc.

Author: Andres Freund
Author: Thomas Munro
Author: David Rowley