Provide smgrreadv() and smgrwritev() for vector I/O.
These map to FileReadV() and FileWriteV(), and thence the OS's vector
I/O routines, after dealing with segmentation.  These functions allow
for contiguous regions of relation files to be transferred into and out
of non-neighboring buffers with a single system call.  The traditional
smgrread() and smgrwrite() functions are implemented in terms of the new
functions.

XXX As for the fd.c functions, the increased likelihood of short
reads/writes (?) is not yet contemplated