refactor lazy_scan_skip and callsite
lazy_scan_heap where it calls lazy_scan_heap had a lot of variables and
it was quite hard to tell what we were actually doing. refactor it a
bit, to make it clear that we are just getting the next block.
lazy_scan_heap() doesn't need to worry about "skippable ranges" and the
like.