Rename LVPagePruneState to PruneResult and relocate to heapam.h
With this rename and relocation we can use PruneResult as an output
parameter for on-access pruning as well. It also makes it harder to
confuse with PruneState.

We'll be adding and removing PruneResult fields in future commits, but
this rename and relocation is separate to make the diff easier to
understand.