Require heap_page_prune() caller to pass non-NULL final argument
Vacuum passes heap_page_prune() an OffsetNumber pointer so that it can
report the offset of the line pointer on which an error was encountered.
The argument was permitted to be NULL, so heap_page_prune() always
tested it before setting it.

On-access pruning (heap_page_prune_opt()), the other caller of
heap_page_prune(), can simply pass a 0 OffsetNumber and avoid the noise
of the extra checks. Do this and assert that a non-NULL argument was
passed in.