Remove old_snapshot_threshold from pruning code
[1] proposes removing old_snapshot_threshold. Based on discussions in
the community, it sounds like this will happen. Removing the
old_snapshot_threshold code from pruning code makes the rest of the
patches in this set simpler to understand. As such, I've added this
throwaway prelim patch to remove the old_snapshot_threshold code which
is in the path of subsequent patches in this set. This patchset has no
hard dependency on old_snapshot threshold being removed, however, so if
it sticks around, I can always drop this patch and fix the others.

[1] postgr.es/m/CACG=ezYV+EvO135fLRdVn-ZusfVsTY6cH1OZqWtezuEYH6ciQA@mail.gmail.com