Remove unneeded post-prune visibility check
heap_page_prune() saves the HTSV for every tuple on a page. By making
this available to its callers, we can avoid an additional call to
HeapTupleSatisfiesVacuum(), saving resources and eliminating the
possibility that vacuuming corrupted data results in a hang due to
endless retry looping.

While we are at it, move some of the output parameters passed into
heap_page_prune() into the PruneResult output parameter, which we need
to pass to heap_page_prune() so we can save htsv in it.