Catalog LP_DEAD items for vacuum while pruning
heap_prune_chain() records which tuples on a page should be marked
unused, dead, or redirected. After returning from heap_page_prune(),
lazy_scan_prune() does a separate loop through all the tuples on the
page and adds the TIDs of the LP_DEAD items to an array. Then it loops
through these TIDs and adds the corresponding items to the array of dead
items maintained by Vacuum. The tuples referred to by these items will
be removed after vacuuming indexes.

Instead, pass a pointer to Vacuum's array of dead items directly to
heap_page_prune() and add the dead items in heap_prune_chain(), at the
time the item is marked LP_DEAD. On-access pruning simply passes NULL
for the dead items argument and nothing is tracked.

This saves vacuum one small loop through the lpdead items' offsets in
lazy_scan_prune(), but, more importantly, it removes one of the
responsibilities of the main loop lazy_scan_prune() does through all the
tuples in the page. Later commits will remove the loop entirely and
combine freezing with pruning.