heap_page_prune() tracks page visibility and counts tuples for vacuum
As soon as we get the result of HeapTupleSatisfiesVacuumHorizon(), we
can count the live and recently dead tuples on the page as well as
determining the visibility_cutoff_xid and determining all_visible for
the page for all non-dead tuples. This was previously done later by
lazy_scan_prune() -- since that information was only used by vacuum.

Setting these counters earlier paves the way for later commits which
will use the all_visible page status and visibility_cutoff_xid during
pruning and move freezing of tuples into pruning. Eventually we will be
able to remove the loop lazy_scan_prune() does through all the tuples on
the page from altogether.

While we are at it, inline heap_prune_satisfies_vacuum() and remove it.
It only had one caller and a few lines of code.