Save a reference to VacuumCutoffs in HeapPageFreeze
We need the relation's VacuumCutoffs whenever we call
heap_prepare_freeze_tuple(), so save a pointer to VacuumCutoffs in
HeapPageFreeze -- a parameter to heap_prepare_freeze_tuple(). This makes
it easier in a future commit to move freezing into pruning without
requiring multiple new arguments to heap_page_prune() that are only used
for freezing.

This is a separate commit for ease of review.